Add removers to InstructorGroup
Added IlmSessionEntity trait and added it to LearnerGroup as well.
