+++ src/Ilios/CoreBundle/Traits/IlmSessionsEntityInterface.php
@@ -1,0 +1,35 @@
+<?php
+
+namespace Ilios\CoreBundle\Traits;
+
+use Doctrine\Common\Collections\ArrayCollection;
+use Doctrine\Common\Collections\Collection;
+
+use Ilios\CoreBundle\Entity\IlmSessionInterface;
+
+/**
+ * Interface IlmSessionsEntityInterface
+ * @package Ilios\CoreBundle\Traits
+ */
+interface IlmSessionsEntityInterface
+{
+    /**
+     * @param Collection $ilmSessions
+     */
+    public function setIlmSessions(Collection $ilmSessions);
+
+    /**
+     * @param IlmSessionInterface $ilmSession
+     */
+    public function addIlmSession(IlmSessionInterface $ilmSession);
+
+    /**
+     * @param IlmSessionInterface $ilmSession
+     */
+    public function removeIlmSession(IlmSessionInterface $ilmSession);
+
+    /**
+    * @return IlmSessionInterface[]|ArrayCollection
+    */
+    public function getIlmSessions();
+}
